+++ src/main/java/lotto/constants/Constants.java
@@ -1,0 +1,9 @@
+package lotto.constants;
+
+public class Constants {
+    public static final int LOTTO_PRICE = 1000;
+    public static final int LOTTO_NUMBER_COUNT = 6;
+    public static final int MIN_LOTTO_NUMBER = 1;
+    public static final int MAX_LOTTO_NUMBER = 45;
+    private static final int[] PRIZES = {0, 2000000000, 30000000, 1500000, 50000, 5000};
+}
