+++ src/main/java/lotto/controller/LottoGame.java
@@ -1,0 +1,64 @@
+package lotto.controller;
+
+import static lotto.constants.Constants.LOTTO_NUMBER_COUNT;
+import static lotto.constants.Constants.MAX_LOTTO_NUMBER;
+import static lotto.constants.Constants.MIN_LOTTO_NUMBER;
+
+import camp.nextstep.edu.missionutils.Randoms;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import lotto.model.Lotto;
+import lotto.model.Rank;
+import lotto.view.InputView;
+import lotto.view.ResultView;
+
+public class LottoGame {
+
+    public void start() {
+        int purchaseAmount = InputView.inputPurchaseAmount();
+        int numberOfLottos = purchaseAmount / 1000;
+        List<Lotto> lottos = generateLottos(numberOfLottos);
+        ResultView.printLottos(lottos);
+
+        List<Integer> winningNumbers = InputView.inputWinningNumbers();
+        int bonusNumber = InputView.inputBonusNumber();
+
+        Map<Rank, Integer> results = calculateResults(lottos, winningNumbers, bonusNumber);
+        ResultView.printResults(results, purchaseAmount);
+    }
+
+    public static List<Lotto> generateLottos(int count) {
+        List<Lotto> lottos = new ArrayList<>();
+        for (int i = 0; i < count; i++) {
+            List<Integer> numbers = Randoms.pickUniqueNumbersInRange(MIN_LOTTO_NUMBER, MAX_LOTTO_NUMBER,
+                    LOTTO_NUMBER_COUNT);
+            Lotto lotto = new Lotto(numbers);
+            lottos.add(lotto);
+        }
+        return lottos;
+    }
+
+    public static Map<Rank, Integer> calculateResults(List<Lotto> lottos, List<Integer> winningNumbers,
+            int bonusNumber) {
+        Map<Rank, Integer> resultMap = new HashMap<>();
+        for (Lotto lotto : lottos) {
+            int matchCount = getMatchCount(lotto.getNumbers(), winningNumbers);
+            boolean bonusMatch = lotto.getNumbers().contains(bonusNumber);
+            Rank rank = Rank.valueOf(matchCount, bonusMatch);
+            resultMap.put(rank, resultMap.getOrDefault(rank, 0) + 1);
+        }
+        return resultMap;
+    }
+
+    public static int getMatchCount(List<Integer> numbers, List<Integer> winningNumbers) {
+        int count = 0;
+        for (int number : numbers) {
+            if (winningNumbers.contains(number)) {
+                count++;
+            }
+        }
+        return count;
+    }
+}
