+++ src/main/java/lotto/validator/Validator.java
@@ -1,0 +1,22 @@
+package lotto.validator;
+
+import static lotto.constants.Constants.LOTTO_PRICE;
+import static lotto.constants.Constants.MAX_LOTTO_NUMBER;
+import static lotto.constants.Constants.MIN_LOTTO_NUMBER;
+
+public class Validator {
+    public static void validatePurchaseAmount(int amount) {
+        if (amount < LOTTO_PRICE) {
+            throw new IllegalArgumentException("[ERROR] 구입 금액은 최소 1,000원 이상이어야 합니다.");
+        }
+        if (amount % LOTTO_PRICE != 0) {
+            throw new IllegalArgumentException("[ERROR] 구입 금액은 1,000원 단위여야 합니다.");
+        }
+    }
+
+    public static void validateNumber(int number) {
+        if (number < MIN_LOTTO_NUMBER || number > MAX_LOTTO_NUMBER) {
+            throw new IllegalArgumentException("[ERROR] 번호는 1부터 45 사이의 숫자여야 합니다.");
+        }
+    }
+}
